requirements.txt
streamlit==1.39.0
trafilatura==1.10.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
pandas==2.2.2
google-generativeai==0.8.3
googlesearch-python==1.0.1